test: 补 trim 后重名测试，去掉多余类型转换
审查指出上一次对 TestCooldownGroupNamesAreTrimmed 的 RED 验证方法有误：
用 stash 实现 + 移出路由得到的 404，只证明了「路由不存在」，没验证
TrimSpace 本身缺失时测试是否会以正确原因失败。已单独做变异测试补证
（临时把 key := strings.TrimSpace(name) 改成 key := name 再跑，确认
失败信息是「实际存了 map[  问候  :5s]」而非 404 或编译错误），验证方法
本身不改代码，不产生新提交内容。

新增 TestCooldownGroupsRejectDuplicateAfterTrim：{"甲":1, "  甲  ":2}
经 trim 后落到同一 key，必须整批拒绝为 422 且不落库，之前只有代码走查
没有测试锁定。

去掉 rule_handler.go 里多余的 int64(maxCooldownMS) 类型转换，
maxCooldownMS 本身已经是 int64。

+++ server/internal/httpapi/rule_handler.go
@@ -308,7 +308,7 @@ func (s *Server) handlePutCooldownGroups(w http.ResponseWriter, r *http.Request)
 		if ms > maxCooldownMS {
 			respondError(w, http.StatusUnprocessableEntity,
 				"冷却组 %s 的间隔 %d 毫秒超过上限（%d 毫秒，即 24 小时）",
-				key, ms, int64(maxCooldownMS))
+				key, ms, maxCooldownMS)
 			return
 		}
 		if _, dup := groups[key]; dup {
